rename object and member file
